Move common code of backup and update page in a new Component container
